Fix array_extensions breaking check for respond_to
When the array_extensions is required it will mess up
expected behavior for expressions like Hash[[]] which should return
an empty hash. This is caused by method_missing logic always producing an empty
Array for any missing method (including :to_hash).

It is assumed that the implementation is there to provide access to the
various 'e' methods. This fix ensures that an empty array is only
produced for methods starting with 'e'.

It may not be accurate (there may be other methods), but it does
at least fix the Hash[[]] problem.