Improve generated setXXX method performance
This commit changes the logic of the generated setXXX method for
many type releationships so that it will create an array where the
contained elements are identity unique. It will then add all elements
without further unique checks.

In essence, this means that the setXXX method uses the method

    Array.uniq {|elem| elem.object_id}

which internally uses a hashed copy to iterate just once. This is a
vast improvement of doing repeated calls to add where each call
performs an increasingly long sequential scan.